fix(converter): type fps as float, since it always accepted one
Nothing enforced the int annotation, and the JS side already does Number(fps) and Math.ceil(animDuration * fps). Fractional rates encode correctly -- 23.976 lands as 2997/125.

Two things the widening needed: _get_animation_output_path interpolates fps into a directory name, so it now formats with %g to keep --video-fps 60 producing _animation_60_png rather than _animation_60.0_png; and pipeline.py wrapped the value in int(), which would have defeated it on that path.

+++ src/mover/converter/render_session.py
@@ -243,7 +243,7 @@ class RenderSession:
         finally:
             self._active_operation = None
 
-    async def get_animation_info(self, fps: int = 60) -> dict[str, Any]:
+    async def get_animation_info(self, fps: float = 60) -> dict[str, Any]:
         page = self._begin_operation("get_animation_info")
         try:
             return await page.evaluate("fps => getAnimationInfo(fps)", fps)
